Navigation Assessment: added nav prefix to components
